Notification services ready to be tested

+++ skillsly_notification_ms/src/infrastructure/adapter/persistence/typeorm/entity/typeorm_notification_change.ts
@@ -25,7 +25,7 @@ export class TypeORMNotificationChange {
 
   @Column({
     name: 'status',
-    type: 'tinyint',
+    type: 'smallint',
     nullable: false
   })
   public status: number;
